+++ react/lib/JBFileInput.tsx
@@ -23,18 +23,18 @@ export const JBFileInput = forwardRef((props:Props, ref) => {
     [element],
   );
   // uploadPercent is passed by ...otherProps
-  const {acceptTypes,initialValue,uploading,label,message,error,disabled,required,maxSize,hideDownload,value,onChange,onInit,onLoad,children, ...otherProps} = props;
-  useJBFileInputAttribute(element,{acceptTypes,initialValue,label,message,error,disabled,required,maxSize,value})
+  const {accept,initialValue,isUploading,label,message,error,disabled,required,maxSize,hideDownload,value,onChange,onInit,onLoad,children, ...otherProps} = props;
+  useJBFileInputAttribute(element,{accept,initialValue,label,message,error,disabled,required,maxSize,isUploading,value})
   useEvents(element,{onChange,onInit,onLoad});
   return (
-    <jb-file-input ref={element} uploading={props.uploading?"":undefined} hide-download={hideDownload?'':undefined} {...otherProps}>{children}</jb-file-input>
+    <jb-file-input ref={element} hide-download={hideDownload?'':undefined} {...otherProps}>{children}</jb-file-input>
   );
 });
 
 JBFileInput.displayName = "JBFileInput";
 type JBFileInputProps = PropsWithChildren<EventProps & JBFileInputAttributes & {
   name?:string,
-  uploading?:boolean,
+  isUploading?:boolean,
   uploadPercent?:number,
   hideDownload?:boolean
 }>